requirements.txt
# RCMaaS - Regulatory Change Monitor as a Service
# Immigration HR Compliance Edition

# Core
anthropic>=0.40.0
python-dotenv>=1.0.0
pyyaml>=6.0.1
schedule>=1.2.0

# Scraping & Fetching
requests>=2.31.0
feedparser>=6.0.10
beautifulsoup4>=4.12.0
lxml>=5.1.0
httpx>=0.27.0

# Content diffing & dedup

# Email (Beehiiv API)
# requests handles this

# Telegram alerts
python-telegram-bot>=21.0

# HTML email templating
jinja2>=3.1.3

# Date/time handling
python-dateutil>=2.9.0

# Logging
loguru>=0.7.2
